% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swcGetMunHistory.R
\name{swcGetMunHistory}
\alias{swcGetMunHistory}
\title{Get municipality mutation history}
\usage{
swcGetMunHistory(munId)
}
\arguments{
\item{munId}{Municipality Id as integer}
}
\description{
The history goes back as far as the mutation number becocmes 1000
  (first time registration)
}
\examples{
library(dplyr)
waedenswil_history <- swcGetMunHistory(293)

}
